Fixed errors caused by porting from windows -> linux + Ported Eigen to glm

+++ src/BunnyPhysicsComponent.h
@@ -12,7 +12,7 @@ public:
 
 	void initObjectPhysics();
 
-	void updatePhysics(double deltaTime);
+	void updatePhysics(float deltaTime);
 
 	void updateBoundingBox();
